+++ views/404.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>My App</title>
+</head>
+<body>
+  <h1>Not Found</h1>
+  <img width="40%" src="https://th.bing.com/th/id/OIG.lYfZrtV9MShHMHjtwD2N?pid=ImgGn" alt="human not found page">
+</body>
+</html>
