+++ views/500.html
@@ -1,0 +1,10 @@
+<html>
+  <head>
+    <link rel="icon" type="image/x-icon" sizes="32x32" href="/favicon.ico">
+    <title>My App</title>
+  </head>
+  <body> 
+    <h1>&#128534; 500 El server esta fuera de servicio</h1>
+    <p>Lo sentimos pero hubo un error en nuestro server...</p>
+  </body>
+</html>
