+++ views/author.html
@@ -1,0 +1,18 @@
+<html>
+
+<head>
+  <link rel="icon" type="image/x-icon" sizes="32x32" href="/favicon.ico">
+  <title>My App</title>
+</head>
+
+<body style="text-align: center;">
+  <h1 style="color: #333;">&#9889; Author &#9889;</h1>
+  <p style="color: #34495E;">Ivan Rivalcoba Rivas - Web Developer</p>
+  <div>
+    <img width="300px"
+      src="https://media.istockphoto.com/id/180841365/photo/hes-a-handsome-man.jpg?s=612x612&w=0&k=20&c=vjQLLI8g_a0O6_xx0plUu3CJ9AMhnSzHssLwgem8gE4="
+      alt="Foto Ivan Rivalcoba">
+  </div>
+</body>
+
+</html>
