+++ views/index.html
@@ -1,0 +1,61 @@
+<html>
+
+<head>
+  <link rel="icon" type="image/x-icon" sizes="32x32" href="/favicon.ico">
+  <title>My App</title>
+  <style>
+    body {
+      background-color: #ECF0F1;
+      font-family: Arial, sans-serif;
+    }
+
+    h1,
+    h2 {
+      color: #3498DB;
+      text-align: center;
+      margin-top: 50px;
+    }
+
+    form {
+      margin-top: 30px;
+      text-align: center;
+    }
+
+    input[type="text"] {
+      width: 300px;
+      padding: 10px;
+      border: none;
+      border-radius: 5px;
+      box-shadow: 0px 0px 5px #3498DB;
+      outline: none;
+    }
+
+    button[type="submit"] {
+      background-color: #3498DB;
+      color: #fff;
+      border: none;
+      border-radius: 5px;
+      padding: 10px 20px;
+      cursor: pointer;
+      box-shadow: 0px 0px 5px #3498DB;
+      outline: none;
+    }
+
+    button[type="submit"]:hover {
+      background-color: #2980B9;
+    }
+  </style>
+</head>
+
+<body>
+  <h1>Hello from my server</h1>
+  <h2>Ingresa un mensaje</h2>
+  <div>
+    <form action="/message" method="POST">
+      <input type="text" name="message">
+      <button type="submit">Send</button>
+    </form>
+  </div>
+</body>
+
+</html>
